backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
mysql-connector-python==9.6.0
bcrypt==5.0.0
scikit-learn==1.5.2
pandas==3.0.1
numpy==2.4.2
python-dotenv==1.2.2
